backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "fastapi>=0.135.3",
    "psycopg2-binary>=2.9.11",
    "psycopg[binary]>=3.3.3",
    "pwdlib[argon2]>=0.3.0",
    "pydantic[email]>=2.12.5",
    "python-dotenv>=1.2.2",
    "python-jose[cryptography]>=3.5.0",
    "sqlmodel>=0.0.38",
    "uvicorn>=0.44.0",
]
